Add surrounding paragraphs
