+++ src/components/SlateEditor/utils/normalizationHelpers.ts
@@ -102,3 +102,36 @@ export const defaultTextBlockNormalizer = (
   }
   nextNormalizeNode(entry);
 };
+
+export const addSurroundingParagraphs = (editor: Editor, path: Path) => {
+  const nextPath = Path.next(path);
+
+  if (Editor.hasPath(editor, nextPath)) {
+    const [nextNode] = Editor.node(editor, nextPath);
+    if (!Element.isElement(nextNode) || !afterOrBeforeTextBlockElement.includes(nextNode.type)) {
+      Transforms.insertNodes(editor, jsx('element', { type: TYPE_PARAGRAPH }), {
+        at: nextPath,
+      });
+
+      return;
+    }
+  }
+
+  if (Path.hasPrevious(path)) {
+    const previousPath = Path.previous(path);
+
+    if (Editor.hasPath(editor, previousPath)) {
+      const [previousNode] = Editor.node(editor, previousPath);
+      if (
+        !Element.isElement(previousNode) ||
+        !afterOrBeforeTextBlockElement.includes(previousNode.type)
+      ) {
+        Transforms.insertNodes(editor, jsx('element', { type: TYPE_PARAGRAPH }), {
+          at: path,
+        });
+
+        return;
+      }
+    }
+  }
+};
